separate standard and 360 formats into individual importers
