+++ lib/dradis/plugins/acunetix.rb
@@ -5,9 +5,10 @@ module Dradis
   end
 end
 
+require 'dradis/plugins/acunetix/acunetix360/importer'
 require 'dradis/plugins/acunetix/engine'
 require 'dradis/plugins/acunetix/field_processor'
-require 'dradis/plugins/acunetix/importer'
 require 'dradis/plugins/acunetix/mapping'
+require 'dradis/plugins/acunetix/standard/importer'
 require 'dradis/plugins/acunetix/version'
